+++ src/instrumentation/llamaindex/patch.ts
@@ -1,29 +1,7 @@
-/*
- * Copyright (c) 2024 Scale3 Labs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 import { LANGTRACE_ADDITIONAL_SPAN_ATTRIBUTES_KEY } from '@langtrace-constants/common'
 import { FrameworkSpanAttributes, Vendors } from '@langtrase/trace-attributes'
-import {
-  Exception,
-  SpanKind,
-  SpanStatusCode,
-  Tracer,
-  context,
-  trace
-} from '@opentelemetry/api'
+import { Tracer, context, SpanKind, SpanStatusCode, Exception, trace } from '@opentelemetry/api'
+import { LangtraceSdkError } from 'errors/sdk_error'
 
 export function genericPatch (
   originalMethod: (...args: any[]) => any,
@@ -33,7 +11,7 @@ export function genericPatch (
   langtraceVersion: string,
   version?: string
 ): (...args: any[]) => any {
-  return async function (this: any, ...args: any[]) {
+  return function (this: any, ...args: any[]) {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-return
     const customAttributes = context.active().getValue(LANGTRACE_ADDITIONAL_SPAN_ATTRIBUTES_KEY) ?? {}
     const attributes: FrameworkSpanAttributes = {
@@ -45,23 +23,48 @@ export function genericPatch (
       'llamaindex.task.name': task,
       ...customAttributes
     }
+    if (task !== 'llamaindex.SimpleVectorStore.query') {
+      attributes['llamaindex.inputs'] = JSON.stringify(args)
+    }
     const spanName = customAttributes['langtrace.span.name' as keyof typeof customAttributes] ?? method
     const span = tracer.startSpan(spanName, { kind: SpanKind.CLIENT, attributes }, context.active())
-    return await context.with(
-      trace.setSpan(context.active(), span),
-      async () => {
-        try {
-          const response = await originalMethod.apply(this, args)
+
+    const invokeMethod = (): any => {
+      try {
+        const result = originalMethod.apply(this, args)
+        if (result instanceof Promise) {
+          // Handle async function
+          return result
+            .then((response: any) => {
+              if (task !== 'llamaindex.SimpleVectorStore.query') {
+                span.setAttributes({ 'llamaindex.outputs': JSON.stringify(response) })
+              }
+              span.setStatus({ code: SpanStatusCode.OK })
+              span.end()
+              return response
+            })
+            .catch((error: any) => {
+              span.recordException(error as Exception)
+              span.setStatus({ code: SpanStatusCode.ERROR })
+              span.end()
+              throw error
+            })
+        } else {
+          // Handle sync function
+          if (task !== 'llamaindex.SimpleVectorStore.query') {
+            span.setAttributes({ 'llamaindex.outputs': JSON.stringify(result) })
+          }
           span.setStatus({ code: SpanStatusCode.OK })
           span.end()
-          return response
-        } catch (error: any) {
-          span.recordException(error as Exception)
-          span.setStatus({ code: SpanStatusCode.ERROR })
-          span.end()
-          throw error
+          return result
         }
+      } catch (error: any) {
+        span.setStatus({ code: SpanStatusCode.ERROR })
+        span.end()
+        throw new LangtraceSdkError(error.message as string, error.stack as string)
       }
-    )
+    }
+
+    return context.with(trace.setSpan(context.active(), span), invokeMethod)
   }
 }