+++ src/instrumentation/vertexai/patch.ts
@@ -40,6 +40,7 @@ import {
   trace,
   Tracer
 } from '@opentelemetry/api'
+import { LangtraceSdkError } from 'errors/sdk_error'
 
 export function generateContentPatch (
   originalMethod: (...args: any[]) => any,
@@ -210,7 +211,7 @@ export function generateContentPatch (
         } catch (error: any) {
           span.setStatus({ code: SpanStatusCode.ERROR })
           span.recordException(error as Exception)
-          throw error
+          throw new LangtraceSdkError(error.message as string, error.stack as string)
         }
       }
     )
@@ -272,7 +273,7 @@ async function * handleStreamResponse (
   } catch (error: any) {
     span.recordException(error as Exception)
     span.setStatus({ code: SpanStatusCode.ERROR })
-    throw error
+    throw new LangtraceSdkError(error.message as string, error.stack as string)
   } finally {
     span.end()
   }